refactor: move todo service to a dedicated package
